fix: commit mistake

+++ Server/src/main/java/dev/zbib/server/repository/MessagesRepository.java
@@ -3,5 +3,5 @@ package dev.zbib.server.repository;
 import dev.zbib.server.model.Message;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MessagesRepository extends JpaRepository<Message, Integer> {
+public interface MessagesRepository extends JpaRepository<Message, Long> {
 }
